+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix broken CLI utilities when a decomp.yaml file is not detected.
+
 ## [2.8.0] - 2025-05-20
 
 ### Added